requirements.txt
python-dotenv~=1.0.0
mpmath~=1.3.0
setuptools~=69.0.3